requirements.txt
fastapi
uvicorn
google-analytics-data
google-auth
gspread
oauth2client
